+++ src/main/java/com/epam/hotel/action/impl/DeleteRoomImageAction.java
@@ -2,19 +2,26 @@ package com.epam.hotel.action.impl;
 
 import com.epam.hotel.action.Action;
 import com.epam.hotel.dao.impl.RoomImageDAOImpl;
+import com.epam.hotel.validation.NumericValidation;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-import static com.epam.hotel.action.impl.ActionConstant.UPLOAD_IMAGE_URL;
+import static com.epam.hotel.action.impl.ActionConstant.*;
 
 public class DeleteRoomImageAction implements Action {
     @Override
     public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-        long imageId = Long.parseLong(request.getParameter("image_id"));
+        if (!NumericValidation.isNumeric(request.getParameter(IMAGE_ID))) {
+            request.setAttribute(MESSAGE, "Изображение не найдено.");
+            request.getRequestDispatcher(ERROR_URL).forward(request, response);
+            return;
+        }
+
+        long imageId = Long.parseLong(request.getParameter(IMAGE_ID));
 
         RoomImageDAOImpl roomImageDAO = new RoomImageDAOImpl();
         roomImageDAO.deleteOneById(imageId);
